Refactor search bar JS components
